float weight ! int

+++ file_operations.py
@@ -3,7 +3,7 @@ def read_file(filename):
         return (
             int(f.readline()),
             int(f.readline()),
-            [int(x) for x in f.readline().split(",")],
+            [float(x) for x in f.readline().split(",")],
             [int(x) for x in f.readline().split(",")],
             [int(x) for x in f.readline().split(",")],
         )
